toolAddMetadataNC

Add metdata nc output files

Usage

toolAddMetadataNC(
  ncFile,
  activityId,
  revision,
  harmonizationPeriod,
  missingValue,
  compression,
  resolution,
  references,
  targetMIP,
  ncTitle,
  referenceDataset,
  furtherInfoUrl
)

Arguments

ncFile

file name of the respective nc file

activityId

a string to store in the nc attribute activityId

revision

a string to store in the nc attribute revision

harmonizationPeriod

a string to store in the nc attribute harmonizationPeriod

missingValue

a string to store in the nc attribute missingValue

compression

a string to store in the nc attribute compression

resolution

a string to store in the nc attribute resolution

references

a string to store in the nc attribute references

targetMIP

a string to store in the nc attribute targetMIP

ncTitle

a string to store in the nc attribute ncTitle

referenceDataset

a string to store in the nc attribute referenceDataset

furtherInfoUrl

a string to store in the nc attribute furtherInfoUrl

Author

Pascal Sauer, Jan Philipp Dietrich